修改贴图和积分规则

+++ Assets/PipeSpawn.cs
@@ -6,7 +6,7 @@ public class PipeSpawn : MonoBehaviour
 {
     // Start is called before the first frame update
     public GameObject pipe;
-    public float spawnRate;
+    public double spawnRate;
     public float timer;
     public float heightOffset;
     void Start()
